requirements.txt
aiosqlite==0.17.0
alembic==1.7.5
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
databases==0.5.4
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.1
ormar==0.10.24
pandas==1.4.0
pkg_resources==0.0.0
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0.post1
email-validator==1.1.3
python-dotenv==0.19.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0